fix(coordinator): track pending transactions without change outputs
