+++ apps/coordinator/src/clients/transactions.ts
@@ -1,32 +1,18 @@
-import { useQuery, useQueryClient, useQueries } from "@tanstack/react-query";
-import { useSelector } from "react-redux";
-import { useMemo, useEffect } from "react";
+import { useQuery, useQueries } from "@tanstack/react-query";
 import { BlockchainClient, TransactionDetails } from "@caravan/clients";
-import {
-  getWalletAddresses,
-  getSpentSlices,
-  Slice,
-  SliceWithLastUsed,
-  WalletState,
-  selectProcessedTransactions,
-} from "selectors/wallet";
-import { calculateTransactionValue } from "utils/transactionCalculations";
+import { Slice } from "selectors/wallet";
 import { useGetClient } from "hooks/client";
 import { bitcoinsToSatoshis } from "@caravan/bitcoin";
-
-const MAX_TRANSACTIONS_TO_FETCH = 500;
 
 // Centralized query key factory for all transaction-related queries
 export const transactionKeys = {
   all: ["transactions"] as const,
   tx: (txid: string) => [...transactionKeys.all, txid] as const,
-  pending: () => [...transactionKeys.all, "pending"] as const,
-  pendingHistory: () => [...transactionKeys.all, "pending", "history"] as const,
   txWithHex: (txid: string) =>
     [...transactionKeys.all, txid, "withHex"] as const,
   coins: (txid: string) => [...transactionKeys.all, txid, "coins"] as const,
-  confirmedHistory: () =>
-    [...transactionKeys.all, "confirmed", "history"] as const,
+  confirmedHistory: () => [...transactionKeys.all, "confirmed"] as const,
+  pendingHistory: () => [...transactionKeys.all, "unconfirmed"] as const,
 };
 
 // Service function for fetching transaction details
@@ -67,130 +53,6 @@ export const useTransactionsWithHex = (txids: string[]) => {
       enabled: !!txid,
     })),
   });
-};
-
-// Basic hook for raw pending transactions (no processing)
-export const usePublicClientPendingTransactions = () => {
-  const blockchainClient = useGetClient();
-  const clientType = blockchainClient?.type;
-  const queryClient = useQueryClient();
-
-  const walletAddresses = useSelector(getWalletAddresses);
-  const spentSlices = useSelector(getSpentSlices) as SliceWithLastUsed[];
-
-  const currentAddresses = useMemo(() => {
-    const spentAddresses =
-      spentSlices
-        ?.map((slice: SliceWithLastUsed) => slice.multisig?.address)
-        .filter(Boolean) || [];
-    return walletAddresses.length > 0 ? walletAddresses : spentAddresses;
-  }, [walletAddresses, spentSlices]);
-
-  useEffect(() => {
-    if (currentAddresses.length > 0 && clientType === "public") {
-      queryClient.invalidateQueries({
-        queryKey: transactionKeys.pendingHistory(),
-      });
-    }
-  }, [currentAddresses.length, clientType, queryClient]);
-
-  return useQuery({
-    queryKey: transactionKeys.pendingHistory(),
-    queryFn: async () => {
-      const rawTransactions =
-        await blockchainClient.getAddressTransactionHistory(
-          currentAddresses,
-          MAX_TRANSACTIONS_TO_FETCH,
-          0,
-        );
-
-      const pendingTx = selectProcessedTransactions(
-        rawTransactions,
-        walletAddresses,
-        "pending",
-      );
-
-      const seenTxids = new Set<string>();
-      const deduplicated = pendingTx.filter((tx) => {
-        if (seenTxids.has(tx.txid)) return false;
-        seenTxids.add(tx.txid);
-        return true;
-      });
-
-      return deduplicated.map((tx) => ({
-        ...tx,
-        valueToWallet: calculateTransactionValue(tx, walletAddresses),
-      }));
-    },
-    enabled:
-      !!blockchainClient &&
-      clientType === "public" &&
-      currentAddresses.length > 0,
-    staleTime: 10000, // 10 seconds
-    refetchInterval: 30000, // 30 seconds
-  });
-};
-
-export const usePrivateClientPendingTransactions = () => {
-  const blockchainClient = useGetClient();
-  const walletAddresses = useSelector(getWalletAddresses);
-  const clientType = blockchainClient?.type;
-  const queryClient = useQueryClient();
-
-  useEffect(() => {
-    if (clientType === "private" && blockchainClient) {
-      queryClient.invalidateQueries({
-        queryKey: transactionKeys.pendingHistory(),
-      });
-    }
-  }, [walletAddresses.length, clientType, queryClient, blockchainClient]);
-
-  return useQuery({
-    queryKey: transactionKeys.pendingHistory(),
-    queryFn: async () => {
-      const rawTransactions =
-        await blockchainClient.getWalletTransactionHistory(
-          MAX_TRANSACTIONS_TO_FETCH,
-          0,
-        );
-
-      const pendingTx = selectProcessedTransactions(
-        rawTransactions,
-        walletAddresses,
-        "pending",
-      );
-
-      const seenTxids = new Set<string>();
-      const deduplicated = pendingTx.filter((tx) => {
-        if (seenTxids.has(tx.txid)) return false;
-        seenTxids.add(tx.txid);
-        return true;
-      });
-
-      return deduplicated.map((tx) => ({
-        ...tx,
-        valueToWallet: calculateTransactionValue(tx, walletAddresses),
-      }));
-    },
-    enabled: !!blockchainClient && clientType === "private",
-    staleTime: 10000,
-    refetchInterval: 30000,
-  });
-};
-
-export const usePendingTransactions = () => {
-  const clientType = useSelector((state: WalletState) => state.client.type);
-  const privateQuery = usePrivateClientPendingTransactions();
-  const publicQuery = usePublicClientPendingTransactions();
-
-  const query = clientType === "private" ? privateQuery : publicQuery;
-
-  return {
-    transactions: query.data || [],
-    isLoading: query.isLoading,
-    error: query.error,
-    refetch: query.refetch,
-  };
 };
 
 export interface Coin {
